pyproject.toml
[tool.poetry]
name = "phidata-testing"
version = "0.1.0"
description = ""
authors = ["slashtechno <77907286+slashtechno@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
phidata = "^2"
duckduckgo-search = "^6.1.4"
ollama = "^0.2.0"
icecream = "^2.1.3"
beautifulsoup4 = "^4.12.3"
exa-py = "^1.0.12"
pgvector = "^0.3.2"
pypdf = "^4.3.1"
psycopg = {extras = ["binary"], version = "^3.2.1"}
sqlalchemy = "^2.0.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
